+++ src/common/config/init.js
@@ -53,6 +53,7 @@ const InitModelFn = () => {
       set allow_registered_children (val) { hiddenState.experimental.allow_registered_children = val },
       duplicate_registered_data: false
     },
+    browser_consent_mode: { enabled: false },
     distributed_tracing: {
       enabled: undefined,
       exclude_newrelic_header: undefined,
@@ -101,7 +102,6 @@ const InitModelFn = () => {
       collect_fonts: false,
       inline_images: false,
       fix_stylesheets: true,
-      // recording config settings
       mask_all_inputs: true,
       // this has a getter/setter to facilitate validation of the selectors
       get mask_text_selector () { return hiddenState.mask_selector },