+++ tests/specs/npm/registered-entity.e2e.js
@@ -90,6 +90,15 @@ describe('registered-entity', () => {
             'newrelic-mfe-id': 2
           }
         })
+        // each payload in this test is decorated with data that matches its appId for ease of testing
+        window.newrelic.recordCustomEvent('CustomEvent', { val: 42 })
+        window.agent1.recordCustomEvent('CustomEvent', { val: 1 })
+        window.agent2.recordCustomEvent('CustomEvent', { val: 2 })
+
+        // each payload in this test is decorated with data that matches its appId for ease of testing
+        window.newrelic.measure('42')
+        window.agent1.measure('1')
+        window.agent2.measure('2')
       })
       const [rumHarvests, errorsHarvests, insightsHarvests, logsHarvest, ajaxHarvest, spaHarvest] = await Promise.all([
         rumCapture.waitForResult({ totalCount: 1, timeout: 10000 }),
@@ -108,9 +117,9 @@ describe('registered-entity', () => {
       // if it gets tried again, the test will fail, since these should all
       // only have one distinct matching payload
       const tests = {
-        42: { rum: false, err: false, pa: false, log: false }, // container agent defaults to appId 42
-        1: { err: false, pa: false, log: false }, // agent1 instance
-        2: { err: false, pa: false, log: false } // agent2 instance
+        42: { rum: false, err: false, pa: false, log: false, rce: false, measure: false }, // container agent defaults to appId 42
+        1: { err: false, pa: false, log: false, rce: false, measure: false }, // agent1 instance
+        2: { err: false, pa: false, log: false, rce: false, measure: false } // agent2 instance
       }
 
       expect(rumHarvests).toHaveLength(1)
@@ -212,10 +221,10 @@ describe('registered-entity', () => {
       errorsHarvests.forEach(({ request: { query, body } }) => {
         const data = body.err
         data.forEach((err, idx) => {
-          const id = err.custom['mfe.id'] || query.a // MFEs use mfe.id, regular agents use appId
+          const id = err.custom['source.id'] || query.a // MFEs use source.id, regular agents use appId
           if (Number(id) !== 42 && testSet.includes('register.jserrors')) {
-            expect(err.custom['mfe.name']).toEqual('agent' + id)
-            expect(err.custom.eventSource).toEqual('MicroFrontendBrowserAgent')
+            expect(err.custom['source.name']).toEqual('agent' + id)
+            expect(err.custom['source.type']).toEqual('MFE')
             expect(err.custom['parent.id']).toEqual(containerAgentEntityGuid)
           } else {
             if (testSet.includes('register') && testSet.includes('register.jserrors')) {
@@ -236,36 +245,56 @@ describe('registered-entity', () => {
       insightsHarvests.forEach(({ request: { query, body } }) => {
         const data = body.ins
         data.forEach((ins, idx) => {
-          if (ins.eventType === 'PageAction') {
-            const id = ins['mfe.id'] || query.a // MFEs use mfe.id, regular agents use appId
+          if (ins.eventType === 'PageAction' || ins.eventType === 'CustomEvent' || (ins.eventType === 'BrowserPerformance' && ins.entryType === 'measure')) {
+            const id = ins['source.id'] || query.a // MFEs use source.id, regular agents use appId
             if (Number(id) !== 42 && testSet.includes('register.generic_events')) {
-              expect(ins['mfe.name']).toEqual('agent' + id)
-              expect(ins.eventSource).toEqual('MicroFrontendBrowserAgent')
+              expect(ins['source.name']).toEqual('agent' + id)
+              expect(ins['source.type']).toEqual('MFE')
               expect(ins['parent.id']).toEqual(containerAgentEntityGuid)
             } else {
               if (testSet.includes('register') && testSet.includes('register.generic_events')) {
                 expect(ins.appId).toEqual(42)
               }
             }
-            countRuns(id, 'pa')
-            if (testSet.includes('register.generic_events')) {
-              expect(ranOnce(id, 'pa')).toEqual(true)
-              expect(Number(id)).toEqual(Number(ins.val))
-            } else {
-              expect(tests[id].pa).toEqual(idx + 1) // each error gets lumped together under the same id without the feature flags
-              expect(Number(id)).toEqual(42)
+            let countType
+            if (ins.eventType === 'PageAction') {
+              countType = 'pa'
+            } else if (ins.eventType === 'CustomEvent') {
+              countType = 'rce'
+            } else if (ins.eventType === 'BrowserPerformance') {
+              countType = 'measure'
             }
+            countRuns(id, countType)
           }
         })
       })
 
+      if (!testSet.includes('register.generic_events')) {
+        // each item gets lumped together under the same id without the feature flags
+        expect(tests['42'].pa).toEqual(testSet.includes('register') ? 3 : 1)
+        expect(tests['42'].rce).toEqual(testSet.includes('register') ? 3 : 1)
+        expect(tests['42'].measure).toEqual(testSet.includes('register') ? 3 : 1)
+      } else {
+        if (testSet.includes('register')) {
+          expect(ranOnce('42', 'pa')).toEqual(true)
+          expect(ranOnce('42', 'rce')).toEqual(true)
+          expect(ranOnce('42', 'measure')).toEqual(true)
+          expect(ranOnce('1', 'pa')).toEqual(true)
+          expect(ranOnce('1', 'rce')).toEqual(true)
+          expect(ranOnce('1', 'measure')).toEqual(true)
+          expect(ranOnce('2', 'pa')).toEqual(true)
+          expect(ranOnce('2', 'rce')).toEqual(true)
+          expect(ranOnce('2', 'measure')).toEqual(true)
+        }
+      }
+
       logsHarvest.forEach(({ request: { query, body } }) => {
         const data = JSON.parse(body)[0]
         data.logs.forEach(log => {
-          const id = log.attributes['mfe.id'] || 42 // MFEs use mfe.id, regular agents supply entity guid so just force it to 42 here if mfe id is not present
+          const id = log.attributes['source.id'] || 42 // MFEs use source.id, regular agents supply entity guid so just force it to 42 here if source id is not present
           if (Number(id) !== 42 && testSet.includes('register')) {
-            expect(log.attributes['mfe.name']).toEqual('agent' + id)
-            expect(log.attributes.eventSource).toEqual('MicroFrontendBrowserAgent')
+            expect(log.attributes['source.name']).toEqual('agent' + id)
+            expect(log.attributes['source.type']).toEqual('MFE')
             expect(log.attributes['parent.id']).toEqual(containerAgentEntityGuid)
           } else {
             if (testSet.includes('register')) {
@@ -308,7 +337,47 @@ describe('registered-entity', () => {
     errorsHarvests.forEach(({ request: { query, body } }) => {
       const data = body.err
       data.forEach(err => {
-        expect(err.custom['mfe.name']).toEqual('agent3')
+        expect(err.custom['source.name']).toEqual('agent2')
+      })
+    })
+  })
+
+  it('should allow a nested register', async () => {
+    await browser.url(await browser.testHandle.assetURL('test-builds/browser-agent-wrapper/registered-entity.html', { init: { feature_flags: ['register', 'register.jserrors'] } }))
+
+    await browser.execute(function () {
+      window.agent1 = new RegisteredEntity({
+        id: 1,
+        name: 'agent1'
+      })
+      window.agent2 = window.agent1.register({
+        id: 2,
+        name: 'agent2'
+      })
+      window.agent3 = window.agent2.register({
+        id: 3,
+        name: 'agent3'
+      })
+      // should get data as "agent2"
+      window.agent1.noticeError('1')
+      window.agent2.noticeError('2')
+      window.agent3.noticeError('3')
+    })
+
+    const errorsHarvests = await mfeErrorsCapture.waitForResult({ totalCount: 1 })
+
+    const containerAgentEntityGuid = await browser.execute(function () {
+      return Object.values(newrelic.initializedAgents)[0].runtime.appMetadata.agents[0].entityGuid
+    })
+
+    // should get ALL data as "agent2" since it replaced the name of agent 1 of the same id
+    errorsHarvests.forEach(({ request: { query, body } }) => {
+      const data = body.err
+      data.forEach((err, idx) => {
+        expect(err.custom['source.name']).toEqual('agent' + (idx + 1))
+        if (idx === 0) expect(err.custom['parent.id']).toEqual(containerAgentEntityGuid) // first app should have container as its parent
+        if (idx === 1) expect(err.custom['parent.id']).toEqual(1) // second app should have first app as its parent
+        if (idx === 2) expect(err.custom['parent.id']).toEqual(2) // third app should have second app as its parent
       })
     })
   })